% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{archs4_cache_experiment_details}
\alias{archs4_cache_experiment_details}
\title{Pull down gse experiment information}
\usage{
archs4_cache_experiment_details(
  x,
  ncores = parallel::detectCores() - 1,
  outdir = NULL,
  ...
)
}
\description{
Pull down gse experiment information
}
