% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{archs4_cache_fn}
\alias{archs4_cache_fn}
\title{File paths for cached objects created from an archs4 hdf5 file}
\usage{
archs4_cache_fn(x, outdir = dirname(x), ...)
}
\arguments{
\item{x}{the path to an archs4 data file}

\item{outdir}{the directory where the cache files are expected to be.
Defaults to \code{dirname(x)}}
}
\description{
File paths for cached objects created from an archs4 hdf5 file
}
\examples{
archs4_cache_fn("~/workspace/data/archs4/v2.6/mouse_gene_v2.6.h5")
}
