% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{archs4_create_cache_files}
\alias{archs4_create_cache_files}
\title{Save the samples and the gene feature info for an archs4 dataset as paruqet
files to make load quicker.}
\usage{
archs4_create_cache_files(x, outdir = dirname(x), ...)
}
\arguments{
\item{x}{the path to the hdf5 file}
}
\description{
Loading the samples table from the hdf5 takes ~25s vs 1s from parquet
}
\examples{
h5fn <- "~/workspace/data/archs4/v2.6/mouse_gene_v2.6.h5"
archs4_create_cache_files(h5fn)
fns <- archs4_cache_fn(h5fn)
sapply(fns, file.exists)
}
