% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{archs4_create_sample_cache}
\alias{archs4_create_sample_cache}
\title{Create sample cache for hdf5 files}
\usage{
archs4_create_sample_cache(
  hdf5 = NULL,
  datadir = getOption("archs4.data_dir"),
  cache_dir = NULL,
  ...,
  create_cache_dir = TRUE
)
}
\arguments{
\item{hdf5}{explicit path to the hdf5 file to create cache for}
}
\description{
You can point to a specific file, or the arcsh4 directory. In the latter
case a cache file will be generated for all hdf5 files that do not have one.
}
