% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arschs4DataSet.R
\name{archs4_creation_date}
\alias{archs4_creation_date}
\title{Return the registered creation date of an arcsh4 dataset}
\usage{
archs4_creation_date(x, string = TRUE, ..., format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{a FacileArchs4DataSet or the path to the source hdf5 data file}

\item{string}{return the data as a string (default) or a \code{Date} object}

\item{...}{stuff and things}

\item{format}{the format return the date in}
}
\value{
the creation date registered in archs4 hdf5 file
}
\description{
Return the registered creation date of an arcsh4 dataset
}
