% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{archs4_dir_info}
\alias{archs4_dir_info}
\title{List the archs4 files in a directory}
\usage{
archs4_dir_info(x, ...)
}
\arguments{
\item{x}{the directory to look inside}
}
\description{
archs4-related files need to have \code{"moue"} or \code{"human"} in their
filename.
}
