% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arschs4DataSet.R
\name{archs4_meta}
\alias{archs4_meta}
\title{Extract metadata from arcsh4 dataset}
\usage{
archs4_meta(x, variable = NULL, ..., multiple_ok = is.character(x))
}
\arguments{
\item{x}{a FacileArchs4DataSet or the path to the source hdf5 data file}

\item{variable}{the name of the variable to extract from the metadata field.
If \code{NULL} (default), all values will be returned.}

\item{...}{stuff and things}

\item{multiple_ok}{If \code{TRUE}, then \code{x} can be a pointer to many hdf5 files}
}
\value{
a list of metadata pulled from the archs4 hdf5 file. If \code{x} is
a list of multiple hdf5 files, then a tibble of metadata is returned,
each row correxponds to the file passed in \code{x}
}
\description{
Extract metadata from arcsh4 dataset
}
