% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arschs4DataSet.R
\name{archs4_sample_search}
\alias{archs4_sample_search}
\title{Find samples that match a regular expression in any of the descriptive
columns.}
\usage{
archs4_sample_search(
  x,
  regex,
  meta_fields = .default_meta_fields(),
  scrub_search = FALSE,
  ignore.case = TRUE,
  remove_sc = x$remove_sc,
  threshold_sc = x$threshold_sc
)
}
\arguments{
\item{x}{A FacileArchs4DataSet or a subset of its sample frame}

\item{regex}{the regular expression to use for search}

\item{meta_fields}{which columns in the dataset to search against, defaults
to a set of columns defined in the archs4py package}

\item{scrub_search}{remove whitespace,dashes,etc. from values in
samples columns before running the regex? Setting this to \code{TRUE} slows
down the search appreciable (~ 3x - 4x). Default: \code{FALSE}}

\item{ignore.case}{Should we ignore case when using \code{regex}?
Default: \code{TRUE}}

\item{remove_sc}{Wether to remove single cell samples from search?
Default: \code{TRUE}}

\item{threshold_sc}{the probability threshold to indicate wheter sample
is from a singlecell dataset, default: 0.5}

\item{...}{stuff and things}
}
\value{
the samples facile_frame that matches the search
}
\description{
This function is fashioned after the archs4py.meta() function.
}
\seealso{
https://github.com/MaayanLab/archs4py/blob/main/archs4py/data.py#L27
}
