% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-utils.R
\name{archs4_split_metadata}
\alias{archs4_split_metadata}
\title{The \code{characteristics_ch1} column often has key:value attributes worth parsing.}
\usage{
archs4_split_metadata(
  x,
  column = "characteristics_ch1",
  delim_var = ",",
  delim_key = ":",
  append = TRUE,
  clean_names = TRUE,
  keep_column = !append,
  ...
)
}
\arguments{
\item{x}{the samples archs4 data.frame}

\item{column}{the name of the column to key:value extract, default
is \code{"characteristics_ch1"}}

\item{delim_var}{the delimeter used to split key:value pairs, default: \code{","}}

\item{delim_key}{the delimeter to split key/value pairs, default: \code{":"}}

\item{append}{if TRUE (default) returns \code{x} with the new columns, otherwise
just returns the metadata}

\item{clean_names}{use \code{\link[janitor:clean_names]{janitor::clean_names()}} to clean up the key variable
names expanded from \code{column}.}

\item{keep_column}{return the \code{x[[column]]} column in the output, defaults
to \code{TRUE}}

\item{...}{stuff and things}
}
\value{
wide tibble of metadata variables from \code{x[[column]]}
}
\description{
Often times the archs4 \code{characteristics_ch1} sample column looks like
\verb{"cell line: K562,cell type: Immortalized epithelial cells,genotype: WT, ...}
Let's bust this up into usable data.frame columns.
}
\details{
TODO: Fix this parsing to accomodate for multiple comma separated values in
on keyvalue par. For instance, the mouse GSE198071 experiment has
characteristics that looks like this:

\code{"tissue: Tibialis anterior muscle,genotype: mfn1-/-, mfn2-/-,age: 6 weeks"}

This gets into the appropriate columns (tissue, genotype, age), however the
genotype column only has mfn1-/- values. The mfn2-/- gets lost
}
\examples{
a4h <- FacileArchs4Data::FacileArchs4DataSet("human")
info.raw <- samples(a4h) |> dplyr::filter(dataset == "GSE212337")
info <- archs4_split_metadata(info.raw)

# This one has values w/ multiple ',' in genotype. I couldn't figure out
# how to encode the tidyr::split_* functions to accomodate for this, which
# is why the key<>value parsing is done inside loops
a4m <- FacileArchs4Data::FacileArchs4DataSet("mouse")
info.raw <- samples(a4m) |> dplyr::filter(dataset == "GSE198071")
info <- archs4_split_metadata(info.raw)
info |> select(h5idx:alignedreads)
table(info$genotype)
}
