% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arschs4DataSet.R
\name{archs4_studies}
\alias{archs4_studies}
\title{Enumerate the studies in an archs4 dataset}
\usage{
archs4_studies(
  x,
  sample_search = NULL,
  meta_fields = .default_meta_fields(),
  threshold_sc = x$threshold_sc,
  threshold_sc_likely = 0.2,
  ...
)
}
\arguments{
\item{x}{the FacileArchs4DataSet}

\item{sample_search}{If a search string is provided, this will trigger
a sample search via the \code{\link[=archs4_sample_search]{archs4_sample_search()}} method. The study table
returned will be augmented with a \code{search_hits} column that tells you how
many samples within the study were a hit for your search.}

\item{threshold_sc}{The singlecellprobability threshold to break for a sample
to be considered likely single cell. This defaults to \code{x$threshold_sc}}

\item{threshold_sc_likely}{the proportion of samples in a study that
must look like singlecell data in order to classify the study as
singlecell. default: 0.2}

\item{scrub_search, meta_fields, ignore.case}{parameters passed to \code{search} function.}
}
\description{
Enumerate the studies in an archs4 dataset
}
