% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arschs4DataSet.R
\name{archs4_version}
\alias{archs4_version}
\title{Return the version of an arcsh4 dataset}
\usage{
archs4_version(x, ...)
}
\arguments{
\item{x}{a FacileArchs4DataSet or the path to the source hdf5 data file}

\item{...}{stuff and things}
}
\value{
the version string from the archs4 hdf5 file
}
\description{
Return the version of an arcsh4 dataset
}
