% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Client.R
\name{Archs4Client}
\alias{Archs4Client}
\title{Thin wrapper to Archs4 dataset}
\description{
Thin wrapper to Archs4 dataset

Thin wrapper to Archs4 dataset
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Archs4Client-new}{\code{Archs4Client$new()}}
\item \href{#method-Archs4Client-clone}{\code{Archs4Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Archs4Client-new"></a>}}
\if{latex}{\out{\hypertarget{method-Archs4Client-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Archs4Client$new(path, ..., sample_columns = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Archs4Client-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Archs4Client-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Archs4Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
