% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-utils.R
\name{arcsh4_split_metadata}
\alias{arcsh4_split_metadata}
\title{The \code{characteristics_ch1} column often has key:value attributes worth parsing.}
\usage{
arcsh4_split_metadata(
  x,
  column = "characteristics_ch1",
  delim_var = ",",
  delim_key = ":",
  clean_names = TRUE,
  keep_column = TRUE,
  ...
)
}
\arguments{
\item{x}{the samples archs4 data.frame}

\item{column}{the name of the column to key:value extract, default
is \code{"characteristics_ch1"}}

\item{delim_var}{the delimeter used to split key:value pairs, default: \code{","}}

\item{delim_key}{the delimeter to split key/value pairs, default: \code{":"}}

\item{clean_names}{use \code{\link[janitor:clean_names]{janitor::clean_names()}} to clean up the key variable
names expanded from \code{column}.}

\item{keep_column}{return the \code{x[[column]]} column in the output, defaults
to \code{TRUE}}

\item{...}{stuff and things}
}
\value{
wide tibble of metadata variables from \code{x[[column]]}
}
\description{
Often times the archs4 \code{characteristics_ch1} sample column looks like
\verb{"cell line: K562,cell type: Immortalized epithelial cells,genotype: WT, ...}
Let's bust this up into usable data.frame columns.
}
\examples{
a4 <- FacileArchs4Data::FacileArchs4DataSet("human")
info.raw <- samples(a4) |> dplyr::filter(dataset == "GSE212337")
info <- arcsh4_split_metadata(info.raw)
}
