% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-Archs4DataSet-facileapi.R
\name{.extract_features}
\alias{.extract_features}
\title{Extract the feature descriptor from the archs4 dataset}
\usage{
.extract_features(x, afds)
}
\description{
Extract the feature descriptor from the archs4 dataset
}
