% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{example_archs4_samples}
\alias{example_archs4_samples}
\title{Return an example sample data frame from a archs4 dataset}
\usage{
example_archs4_samples(species = c("mouse", "human"), afds = NULL, ...)
}
\description{
Return an example sample data frame from a archs4 dataset
}
