% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arschs4DataSet.R
\name{FacileArchs4DataSet}
\alias{FacileArchs4DataSet}
\title{Constructor}
\usage{
FacileArchs4DataSet(
  x,
  ...,
  species = NULL,
  data_dir = getOption("archs4.data_dir", NULL),
  use_cache = TRUE,
  remove_sc = TRUE,
  threshold_sc = 0.5
)
}
\arguments{
\item{x}{name of species, of path to the h5 file}
}
\description{
Constructor
}
\examples{
mfds <- FacileArchs4DataSet("mouse")

}
