% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile-api.R
\name{fetch_assay_data.archs4_client_facile_frame}
\alias{fetch_assay_data.archs4_client_facile_frame}
\title{Retrieve assay data}
\usage{
\method{fetch_assay_data}{archs4_client_facile_frame}(
  x,
  features = NULL,
  samples = NULL,
  assay_name = NULL,
  normalized = FALSE,
  batch = NULL,
  main = NULL,
  as.matrix = FALSE,
  drop_samples = TRUE,
  ...,
  prior.count = 3,
  subset.threshold = 700,
  aggregate = FALSE,
  aggregate.by = "ewm",
  verbose = FALSE
)
}
\description{
Retrieve assay data
}
