% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileapi-S3.R
\name{fetch_assay_data.FacileArchs4DataSet}
\alias{fetch_assay_data.FacileArchs4DataSet}
\title{Retrieve assay data}
\usage{
\method{fetch_assay_data}{FacileArchs4DataSet}(
  x,
  features = NULL,
  samples = NULL,
  assay_name = NULL,
  normalized = FALSE,
  batch = NULL,
  main = NULL,
  as.matrix = FALSE,
  drop_samples = TRUE,
  ...,
  prior.count = 0.25,
  subset.threshold = 700,
  aggregate = FALSE,
  aggregate.by = "ewm",
  verbose = FALSE
)
}
\description{
Retrieve assay data
}
\examples{
m4 <- FacileArchs4DataSet()
xf <- features(m4) |> filter(name \%in\% c("Sesn2", "Fgf21"))
xs <- samples(m4) |> filter(grepl("GSE249764", dataset))
}
