% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_search.R
\name{fuzzy_search}
\alias{fuzzy_search}
\title{Fuzzy search across columns to find samples that match search term}
\usage{
fuzzy_search(
  x,
  regex,
  meta_fields = NULL,
  scrub_search = FALSE,
  ignore.case = TRUE,
  remove_sc = TRUE,
  threshold_sc = 0.5
)
}
\arguments{
\item{x}{a tibble of sample information}

\item{regex}{the regular expression to use for search}

\item{meta_fields}{which columns in the dataset to search against, defaults
to a set of columns defined in the archs4py package}

\item{scrub_search}{remove whitespace,dashes,etc. from values in
samples columns before running the regex? Setting this to \code{TRUE} slows
down the search appreciable (~ 3x - 4x). Default: \code{FALSE}}

\item{ignore.case}{Should we ignore case when using \code{regex}?
Default: \code{TRUE}}

\item{remove_sc}{Wether to remove single cell samples from search?
Default: \code{TRUE}}

\item{threshold_sc}{the probability threshold to indicate wheter sample
is from a singlecell dataset, default: 0.5}

\item{...}{stuff and things}
}
\value{
the samples tibble that matches the search
}
\description{
Search for samples that are "kind of like" soething you are interested in
}
