% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileapi-R6.R, R/facileapi-S3.R
\name{with_assay_data.archs4_client_facile_frame}
\alias{with_assay_data.archs4_client_facile_frame}
\title{Retrieve (wide) assay data}
\usage{
\method{with_assay_data}{archs4_client_facile_frame}(
  x,
  features,
  assay_name = NULL,
  normalized = TRUE,
  aggregate = FALSE,
  aggregate.by = "ewm",
  spread = TRUE,
  with_assay_name = FALSE,
  ...,
  verbose = FALSE,
  .fds = fds(x)
)

\method{with_assay_data}{archs4_client_facile_frame}(
  x,
  features,
  assay_name = NULL,
  normalized = TRUE,
  aggregate = FALSE,
  aggregate.by = "ewm",
  spread = TRUE,
  with_assay_name = FALSE,
  ...,
  verbose = FALSE,
  .fds = fds(x)
)
}
\description{
Retrieve (wide) assay data

Retrieve (wide) assay data
}
